feat: add playground components and hooks for number, text, select fields, and toggle
- Implemented `NumberField`, `TextField`, `SelectField`, and `Toggle` components with respective hooks.
- Introduced `useImperativeApi` for managing imperative API interactions.
- Created `StatePanel` for displaying Swipi state and positions.
- Developed `Stage` component to render slides with Swipi integration.
- Added constants and helper functions for playground state management.
- Implemented `usePlayground` hook for managing playground state and events.
- Added CSS styles for new components and overall playground layout.
