+++ index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="src/dev/main.tsx"></script>
+    <script type="module" src="src/Playground/main.tsx"></script>
   </body>
 </html>
