+++ src/Playground/components/ConfigEditor/constants.ts
@@ -1,0 +1,16 @@
+import { ConfigType } from '../../../Swipi/types'
+import { ConfigNumberFieldOption } from '../../types'
+
+export const CONFIG_NUMBER_FIELDS: ConfigNumberFieldOption[] = [
+  { key: 'maxWidth', label: 'maxWidth' },
+  { key: 'slidesNumber', label: 'slidesNumber' },
+  { key: 'spaceBetween', label: 'spaceBetween' }
+]
+
+export const NEW_CONFIG_ITEM: ConfigType = {
+  maxWidth: 640,
+  slidesNumber: 1,
+  spaceBetween: 10
+}
+
+export const EMPTY_FIELD_VALUE = 0
