+++ src/Playground/components/ControlsPanel/constants.ts
@@ -1,0 +1,11 @@
+export const SLIDES_NUMBER_LIMITS = { min: 1, max: 6 }
+
+export const SPACE_BETWEEN_LIMITS = { min: 0, max: 80 }
+
+export const AUTOPLAY_SPEED_LIMITS = { min: 500, max: 10000, step: 100 }
+
+export const ANIMATION_SPEED_LIMITS = { min: 0, max: 2000, step: 50 }
+
+export const DOT_SIZE_LIMITS = { min: 4, max: 40 }
+
+export const STAGE_WIDTH_LIMITS = { min: 240, max: 1440, step: 10 }
