+++ src/Playground/components/ControlsPanel/useControlsPanel.ts
@@ -1,0 +1,20 @@
+import {
+  PlaygroundState,
+  PlaygroundStateKey,
+  UseControlsPanelProps,
+  UseControlsPanelReturn
+} from '../../types'
+
+export const useControlsPanel = ({
+  update
+}: UseControlsPanelProps): UseControlsPanelReturn => {
+  const change =
+    <Key extends PlaygroundStateKey>(key: Key) =>
+    (value: PlaygroundState[Key]): void =>
+      update(key, value)
+
+  return {
+    change,
+    changeStageWidth: (width: number) => () => update('stageWidth', width)
+  }
+}
