+++ src/Playground/components/CustomActiveDot/index.tsx
@@ -1,0 +1,7 @@
+import type { JSX } from 'react'
+
+const CustomActiveDot = (): JSX.Element => (
+  <span className="pg-custom-dot pg-custom-dot--active" />
+)
+
+export default CustomActiveDot
