+++ src/Playground/components/CustomDot/index.tsx
@@ -1,0 +1,5 @@
+import type { JSX } from 'react'
+
+const CustomDot = (): JSX.Element => <span className="pg-custom-dot" />
+
+export default CustomDot
