+++ src/Playground/components/ImperativeApi/constants.ts
@@ -1,0 +1,3 @@
+export const FIRST_INDEX = 0
+
+export const LAST_INDEX_OFFSET = 1
