+++ src/Playground/components/NumberField/constants.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_MIN = 0
+
+export const DEFAULT_MAX = 100
+
+export const DEFAULT_STEP = 1
