+++ src/Playground/components/NumberField/helpers.ts
@@ -1,0 +1,2 @@
+export const clamp = (value: number, min: number, max: number): number =>
+  Math.min(Math.max(value, min), max)
