+++ src/Playground/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import Playground from './Playground'
+import Playground from '.'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
